refactor: move similar code on wallet creation and import to mixin
